+++ lib/elisper/environment/list_env.ex
@@ -6,6 +6,7 @@ defmodule Elisper.Environment.ListEnv do
   resembles the implementation in SICP using linked lists.
   """
   alias Elisper.Environment.UnboundVariableError
+  alias Elisper.Environment.ListFrame, as: Frame
 
   @doc """
   Get an empty environment.
@@ -19,7 +20,7 @@ defmodule Elisper.Environment.ListEnv do
 
   def lookup_variable_value(variable, env) do
     frame = first_frame(env)
-    result = find_in_frame(variable, frame)
+    result = Frame.find_in_frame(variable, frame)
 
     case result do
       nil -> lookup_variable_value(variable, enclosing_environment(env))
@@ -38,7 +39,7 @@ defmodule Elisper.Environment.ListEnv do
 
   def set_variable_value(variable, value, env) do
     [frame, enclosing] = env
-    result = find_in_frame(variable, frame)
+    result = Frame.find_in_frame(variable, frame)
 
     case result do
       nil ->
@@ -46,7 +47,7 @@ defmodule Elisper.Environment.ListEnv do
         {result, [frame, env]}
 
       _ ->
-        updated_frame = replace_in_frame(variable, value, frame)
+        updated_frame = Frame.replace_in_frame(variable, value, frame)
         {:ok, [updated_frame, enclosing]}
     end
   end
@@ -58,7 +59,7 @@ defmodule Elisper.Environment.ListEnv do
     do: raise(UnboundVariableError)
 
   def extend_environment(variables, values, env),
-    do: [make_frame(variables, values), env]
+    do: [Frame.make_frame(variables, values), env]
 
   @doc """
   Defines a new variable or update if variable already bound
@@ -67,72 +68,12 @@ defmodule Elisper.Environment.ListEnv do
     {result, env} = set_variable_value(variable, value, env)
 
     case result do
-      :unbound -> [add_binding_to_frame(variable, value, first_frame), enclosing]
+      :unbound -> [Frame.add_binding_to_frame(variable, value, first_frame), enclosing]
       :ok -> env
     end
   end
 
-  @doc """
-    Gets the "enclosing" environment consisting of all trailing frames.
-  """
-  def enclosing_environment([_, enclosing]), do: enclosing
+  defp enclosing_environment([_, enclosing]), do: enclosing
+  defp first_frame([first_frame, _]), do: first_frame
 
-  @doc """
-    Gets the first frame in the supplied environment.
-  """
-  def first_frame([first_frame, _]), do: first_frame
-
-  @doc """
-  Makes a new frame given a list varaibles and a list of values to be bound.
-  """
-  def make_frame(variables, values), do: [variables, values]
-
-  @doc """
-  Gets the list of variables in the specified frame.
-  """
-  def frame_variables([variables, _]), do: variables
-
-  @doc """
-  Gets the list of bound values in the specified frame.
-  """
-  def frame_values([_, values]), do: values
-
-  @doc """
-  Adds a new variable binding to a frame.
-  """
-  def add_binding_to_frame(variable, value, []), do: [[variable], [value]]
-
-  def add_binding_to_frame(variable, value, frame) do
-    [[variable] ++ frame_variables(frame), [value] ++ frame_values(frame)]
-  end
-
-  @doc """
-  Tries to find the value bound to specified variable in the supplied frame.
-  """
-  def find_in_frame(var, frame) do
-    vars = frame_variables(frame)
-    idx = Enum.find_index(vars, fn item -> item == var end)
-
-    if idx != nil do
-      {:ok, val} = Enum.fetch(frame_values(frame), idx)
-      val
-    end
-  end
-
-  @doc """
-  Tries to replace a variable value in the supplied frame.
-  """
-  def replace_in_frame(var, val, [vars, vals]) do
-    tframe =
-      Enum.zip(vars, vals)
-      |> Enum.map(fn binding ->
-        cond do
-          elem(binding, 0) == var -> {var, val}
-          true -> binding
-        end
-      end)
-      |> Enum.unzip()
-
-    [elem(tframe, 0), elem(tframe, 1)]
-  end
 end
