+++ test/environment/list_env_test.exs
@@ -8,35 +8,6 @@ defmodule ListEnvTest do
     assert ListEnv.empty() == []
   end
 
-  test "make_frame builds a pair of lists" do
-    variables = ["x", "y", "z"]
-    values = [1, 2, 3]
-    expected_frame = [["x", "y", "z"], [1, 2, 3]]
-    frame = ListEnv.make_frame(variables, values)
-    assert frame == expected_frame
-  end
-
-  test "frame_variables returns variables" do
-    expected_variables = ["x", "y", "z"]
-    frame = [["x", "y", "z"], [1, 2, 3]]
-    variables = ListEnv.frame_variables(frame)
-    assert variables == expected_variables
-  end
-
-  test "frame_values returns values" do
-    expected_values = [1, 2, 3]
-    frame = [["x", "y", "z"], [1, 2, 3]]
-    values = ListEnv.frame_values(frame)
-    assert values == expected_values
-  end
-
-  test "add binding adds new binding to frame" do
-    frame = [["x", "y", "z"], [1, 2, 3]]
-    expected_frame = [["foo", "x", "y", "z"], ["bar", 1, 2, 3]]
-    new_frame = ListEnv.add_binding_to_frame("foo", "bar", frame)
-    assert new_frame == expected_frame
-  end
-
   test "extend environment appends new frame to current environment" do
     base_environment = []
     vars = ["x"]
@@ -51,30 +22,6 @@ defmodule ListEnvTest do
     vars = ["x", "y"]
     vals = [1]
     assert_raise UnboundVariableError, fn -> ListEnv.extend_environment(vars, vals, []) end
-  end
-
-  test "find_in_frame finds value when in frame" do
-    frame = [["x", "y", "z"], [1, 2, 3]]
-    assert ListEnv.find_in_frame("y", frame) == 2
-  end
-
-  test "find_in_frame returns nil when not in frame" do
-    frame = [["x", "y", "z"], [1, 2, 3]]
-    assert ListEnv.find_in_frame("q", frame) == nil
-  end
-
-  test "first_frame returns first frame" do
-    first_frame = [["x"], [1]]
-    enclosing = [["y"], [2]]
-    env = [first_frame, enclosing]
-    assert ListEnv.first_frame(env) == first_frame
-  end
-
-  test "enclosing_environment returns enclosing" do
-    first_frame = [["x"], [1]]
-    enclosing = [["y"], [2]]
-    env = [first_frame, enclosing]
-    assert ListEnv.enclosing_environment(env) == enclosing
   end
 
   test "lookup variable raises unbound variable error when empty env" do
@@ -110,12 +57,6 @@ defmodule ListEnvTest do
     assert updated_env == env
   end
 
-  test "replace in frame returns updated frame" do
-    frame = [["x", "y", "z"], [1, 2, 3]]
-    expected = [["x", "y", "z"], [1, 17, 3]]
-    assert ListEnv.replace_in_frame("y", 17, frame) == expected
-  end
-
   test "set variable value returns updated environment when success" do
     var = "x"
     val = 2
@@ -127,14 +68,6 @@ defmodule ListEnvTest do
     {result, updated_env} = ListEnv.set_variable_value(var, val, env)
     assert updated_env == expected_env
     assert result == :ok
-  end
-
-  test "add_binding_to_frame adds new binding" do
-    var = "x"
-    val = 3.14
-    frame = [["hello"], ["world"]]
-    expected_frame = [["x", "hello"], [3.14, "world"]]
-    assert ListEnv.add_binding_to_frame(var, val, frame) == expected_frame
   end
 
   test "define variable updates variable if vaiable already bound" do
